email_summarizer/requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
transformers
torch
sentencepiece
